backend/requirements.txt
fastapi
uvicorn
pydantic
python-keycloak
python-jose[cryptography]
